requirements.txt
Flask==2.1.0
gunicorn==20.1.0
Werkzeug==2.3.7
deeplake
langchain
openai
chromadb
tiktoken
GitPython
sentence_transformers
streamlit
python-dotenv
slack-bolt